Moving clone course mutatino to CloneModal component
