+++ frontend/src/components/Feed/CloneModal.tsx
@@ -12,11 +12,15 @@ import {
   ModalFooter,
   Button,
   Text,
+  useToast,
 } from "@chakra-ui/react";
-import axios from "axios";
+import axios, { AxiosError } from "axios";
 import { formatDistanceToNow } from "date-fns";
 import React from "react";
-import { useQuery } from "react-query";
+import { useMutation, useQuery } from "react-query";
+import { useNavigate } from "react-router-dom";
+import { ICourse } from "../../types/course";
+import { ErrorType } from "../../types/requests";
 import { IResource } from "../../types/resource";
 import { getConfig, useSession } from "../../utils/auth";
 import { baseURL } from "../../utils/constants";
@@ -29,6 +33,10 @@ interface Props {
   courseCreatedAt: string;
   courseId: string;
 }
+interface CloneForm {
+  name: string;
+  code: string;
+}
 
 const CloneModal = ({
   isOpen,
@@ -38,12 +46,10 @@ const CloneModal = ({
   courseCreatedAt,
 }: Props) => {
   const { user } = useSession();
+  const toast = useToast();
+  const navigate = useNavigate();
 
-  const {
-    data: resources,
-    isFetching,
-    isLoading,
-  } = useQuery(
+  const { data: resources, isLoading } = useQuery(
     `hub-${courseId}`,
     async () => {
       if (!user) throw new Error("User is not defined");
@@ -59,6 +65,48 @@ const CloneModal = ({
       enabled: isOpen && !!user && !!courseId,
     }
   );
+
+  const cloneCourse = useMutation(
+    async ({ name, code }: CloneForm) => {
+      if (!user) {
+        throw new Error("User not defined");
+        return;
+      }
+      const res = await axios.post<ICourse>(
+        `${baseURL}/course/clone/${code}`,
+        {
+          name,
+        },
+        getConfig(user?.sessionToken)
+      );
+      return res.data;
+    },
+    {
+      onSuccess: (course) => {
+        toast({
+          status: "success",
+          title: "Course cloned!",
+          description: "The course has successfully been cloned",
+          isClosable: true,
+        });
+        navigate(`/courses/${course?.objectId}/difficulty`);
+      },
+      onError: (error: AxiosError<ErrorType>) => {
+        toast({
+          title: "An error ocurred",
+          description: error.response?.data.message,
+          status: "error",
+          duration: 5000,
+          isClosable: true,
+        });
+      },
+    }
+  );
+
+  const handleCloneCourse = () => {
+    if (!courseId || !courseName) return;
+    cloneCourse.mutate({ code: courseId, name: courseName });
+  };
 
   return (
     <Modal isOpen={isOpen} onClose={onClose} size="4xl">
@@ -98,7 +146,13 @@ const CloneModal = ({
           <Button onClick={onClose} mr={3}>
             Cancel{" "}
           </Button>
-          <Button colorScheme="green">Clone Course</Button>
+          <Button
+            colorScheme="green"
+            onClick={() => handleCloneCourse()}
+            isLoading={cloneCourse.isLoading}
+          >
+            Clone Course
+          </Button>
         </ModalFooter>
       </ModalContent>
     </Modal>
